+++ src/wrapt/proxies.py
@@ -295,7 +295,7 @@ class LazyObjectProxy(AutoObjectProxy):
 
         super().__init__(None)
 
-    __wrapped_initialized__ = False
+    __wrapped_get_called__ = False
 
     def __wrapped_factory__(self):
         return None
@@ -315,12 +315,12 @@ class LazyObjectProxy(AutoObjectProxy):
             # If it is then just return it, otherwise call the factory to
             # create it.
 
-            if self.__wrapped_initialized__:
+            if self.__wrapped_get_called__:
                 return self.__wrapped__
 
             self.__wrapped__ = self.__wrapped_factory__()
 
-            self.__wrapped_initialized__ = True
+            self.__wrapped_get_called__ = True
 
             return self.__wrapped__